# Optimize dev edit-refresh loop by caching the highlighted files