--- a/packages/lit-dev-tools-cjs/src/playground-plugin/blocking-renderer.ts
+++ b/packages/lit-dev-tools-cjs/src/playground-plugin/blocking-renderer.ts
@@ -6,6 +6,12 @@
 
 import * as workerthreads from 'worker_threads';
 import * as pathlib from 'path';
+import * as fs from 'fs';
+
+const cachedHighlightsDir = pathlib.resolve(
+  __dirname,
+  '../../.highlights_cache/'
+);
 
 export type WorkerMessage = HandshakeMessage | Render | Shutdown;
 
@@ -32,6 +38,33 @@ export interface Shutdown {
   type: 'shutdown';
 }
 
+const highlightKey = (lang: string, code: string) => `[${lang}]:${code}`;
+
+// Create a cache key for the highlighted strings. This is a
+// simple digest build from a DJB2-ish hash modified from:
+// https://github.com/darkskyapp/string-hash/blob/master/index.js
+// It is modified from @lit-labs/ssr-client.
+// Goals:
+//  - Extremely low collision rate. We may not be able to detect collisions.
+//  - Extremely fast.
+//  - Extremely small code size.
+//  - Safe to include in HTML comment text or attribute value.
+//  - Easily specifiable and implementable in multiple languages.
+// We don't care about cryptographic suitability.
+export const digestToFileName = (stringToDigest: string) => {
+  // Number of 32 bit elements to use to create template digests
+  const digestSize = 5;
+  const hashes = new Uint32Array(digestSize).fill(5381);
+  for (let i = 0; i < stringToDigest.length; i++) {
+    hashes[i % digestSize] =
+      (hashes[i % digestSize] * 33) ^ stringToDigest.charCodeAt(i);
+  }
+  const str = String.fromCharCode(...new Uint8Array(hashes.buffer));
+  return Buffer.from(str, 'binary')
+    .toString('base64')
+    .replace(/[<>:"'/\\|?*]/g, '_');
+};
+
 export class BlockingRenderer {
   /** Worker that performs rendering. */
   private worker: workerthreads.Worker;
@@ -44,8 +77,14 @@ export class BlockingRenderer {
   private decoder = new TextDecoder();
   private exited = false;
   private renderTimeout: number;
+  private isDevMode = false;
 
-  constructor({renderTimeout = 60_000, maxHtmlBytes = 1024 * 1024} = {}) {
+  constructor({
+    renderTimeout = 60_000,
+    maxHtmlBytes = 1024 * 1024,
+    isDevMode = false,
+  } = {}) {
+    this.isDevMode = isDevMode;
     this.renderTimeout = renderTimeout;
     this.sharedHtml = new Uint8Array(new SharedArrayBuffer(maxHtmlBytes));
     this.worker = new workerthreads.Worker(
@@ -70,6 +109,11 @@ export class BlockingRenderer {
       htmlBuffer: this.sharedHtml,
       notify: this.sharedNotify,
     });
+    try {
+      fs.mkdirSync(cachedHighlightsDir);
+    } catch {
+      // Directory already exists.
+    }
   }
 
   async stop(): Promise<void> {
@@ -82,9 +126,35 @@ export class BlockingRenderer {
     });
   }
 
+  private getCachedRender(lang: string, code: string): string | null {
+    if (!this.isDevMode) {
+      return null;
+    }
+    const fileName = digestToFileName(highlightKey(lang, code));
+    const absoluteFilePath = pathlib.resolve(cachedHighlightsDir, fileName);
+    if (fs.existsSync(absoluteFilePath)) {
+      return fs.readFileSync(absoluteFilePath, {encoding: 'utf8'});
+    }
+    return null;
+  }
+
+  private writeCachedRender(lang: string, code: string, html: string) {
+    if (!this.isDevMode) {
+      // In production mode, don't write out cached files.
+      return;
+    }
+    const fileName = digestToFileName(highlightKey(lang, code));
+    const absoluteFilePath = pathlib.resolve(cachedHighlightsDir, fileName);
+    fs.writeFileSync(absoluteFilePath, html);
+  }
+
   render(lang: 'js' | 'ts' | 'html' | 'css', code: string): {html: string} {
     if (this.exited) {
       throw new Error('BlockingRenderer worker has already exited');
+    }
+    const cachedResult = this.getCachedRender(lang, code);
+    if (cachedResult !== null) {
+      return {html: cachedResult};
     }
     this.workerPost({type: 'render', lang, code});
     if (
@@ -97,6 +167,7 @@ export class BlockingRenderer {
     const raw = this.decoder.decode(this.sharedHtml);
     const length = this.sharedLength[0];
     const html = raw.substring(0, length);
+    this.writeCachedRender(lang, code, html);
     return {html};
   }