Issue: https://github.com/lit/lit.dev/issues/1201

### Context

This PR focusses exclusively on faster edit-refresh dev mode times.

Result is 11.3s edit-refresh time (from 16.7s edit-refresh time) on my M2 silicon mac.
Or, 34.4s edit-refresh time (from 57.3s ) on intel mac.

### How

Instead of using the playwright browser process to always retrieve the syntax highlighted code, instead use a temporary dev cache so syntax highlighted code blocks can be retrieved from the file system.

This ends up being more optimal than running the code through the browser (even using synchronous file system calls).

Note: synchronous file system calls must be used as the whole API must be sync for 11ty.

### Test plan

Manually tried it with edit refresh.
Note, there is a risk of a collision. In this case the code may appear wrong **only in dev mode** where caching is enabled.